trees: Default to using float datatype
Start of making it configurable through convert()

+++ emlearn/convert.py
@@ -3,15 +3,15 @@ from . import trees
 from . import net
 from . import bayes
 
-def convert(estimator, kind=None, method='pymodule'):
+def convert(estimator, kind=None, method='pymodule', dtype='float'):
     """Main entrypoint for converting a model"""
 
     if kind is None:
         kind = type(estimator).__name__
 
-    # Uname instead of instance to avoid hard dependency on the libraries
+    # Use name instead of instance to avoid hard dependency on the libraries
     if kind in ['RandomForestClassifier', 'ExtraTreesClassifier', 'DecisionTreeClassifier']:
-        return trees.Wrapper(estimator, method) 
+        return trees.Wrapper(estimator, method, dtype=dtype)
     elif kind == 'MLPClassifier':
         return net.convert_sklearn_mlp(estimator, method)
     elif kind == 'Sequential':
